scripts/requirements.txt
tqdm
playwright
PyMuPDF
boto3
python-dotenv
